+++ Modules/Intent.Modules.Modelers.Services/Api/CommentModel.cs
@@ -1,0 +1,82 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
+using Intent.Modules.Common;
+using Intent.Modules.Common.Types.Api;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementModel", Version = "1.0")]
+
+namespace Intent.Modelers.Services.Api
+{
+    [IntentManaged(Mode.Fully, Signature = Mode.Fully)]
+    public partial class CommentModel : IMetadataModel, IHasStereotypes, IHasName, IElementWrapper, IHasFolder
+    {
+        public const string SpecializationType = "Comment";
+        public const string SpecializationTypeId = "32cb9020-2896-4dc0-9a6d-2aaae3cb431f";
+        protected readonly IElement _element;
+
+        [IntentManaged(Mode.Fully)]
+        public CommentModel(IElement element, string requiredType = SpecializationType)
+        {
+            if (!requiredType.Equals(element.SpecializationType, StringComparison.InvariantCultureIgnoreCase))
+            {
+                throw new Exception($"Cannot create a '{GetType().Name}' from element with specialization type '{element.SpecializationType}'. Must be of type '{SpecializationType}'");
+            }
+            _element = element;
+            Folder = _element.ParentElement?.SpecializationTypeId == FolderModel.SpecializationTypeId ? new FolderModel(_element.ParentElement) : null;
+        }
+
+        public string Id => _element.Id;
+
+        public string Name => _element.Name;
+
+        public string Comment => _element.Comment;
+
+        public IEnumerable<IStereotype> Stereotypes => _element.Stereotypes;
+
+        public FolderModel Folder { get; }
+
+        public IElement InternalElement => _element;
+
+        public override string ToString()
+        {
+            return _element.ToString();
+        }
+
+        public bool Equals(CommentModel other)
+        {
+            return Equals(_element, other?._element);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(this, obj)) return true;
+            if (obj.GetType() != this.GetType()) return false;
+            return Equals((CommentModel)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            return (_element != null ? _element.GetHashCode() : 0);
+        }
+    }
+
+    [IntentManaged(Mode.Fully)]
+    public static class CommentModelExtensions
+    {
+
+        public static bool IsCommentModel(this ICanBeReferencedType type)
+        {
+            return type != null && type is IElement element && element.SpecializationTypeId == CommentModel.SpecializationTypeId;
+        }
+
+        public static CommentModel AsCommentModel(this ICanBeReferencedType type)
+        {
+            return type.IsCommentModel() ? new CommentModel((IElement)type) : null;
+        }
+    }
+}