+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -2,6 +2,10 @@
 
 - Improvement: Added support for Razor file code management.
 
+### Version 3.7.2
+
+- Improvement: Added Attribute support for Class Constructor parameters.
+
 ### Version 3.7.1
 
 - Improvement: Support for `ValueTask` as a `CSharpType` on Methods.