+++ Modules/Intent.Modules.Modelers.Services/release-notes.md
@@ -1,3 +1,7 @@
+### Version 3.7.5
+
+- Improvement: It is now possible to create a _Comment_ element type in the _Services_ designer. Inline with comments which already existed in the _Domain_ designer, these comments can be placed on a diagram and associated with other elements.
+
 ### Version 3.7.4
 
 - Fixed: Ensure that when the `Web Client` module is installed that it will install the Service designer package reference.